requirements.txt
# Core RAG Orchestration
llama-index>=0.10.0
llama-index-core>=0.10.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0

# Vector Database
qdrant-client>=1.7.0
llama-index-vector-stores-qdrant>=0.1.0

# Observability
arize-phoenix>=4.0.0
openinference-instrumentation-llama-index>=2.0.0

# User Interface
streamlit>=1.30.0

# Evaluation
ragas>=0.1.0
pandas>=2.0.0
datasets>=2.14.0

# Document Processing
pypdf>=4.0.0
markdown>=3.5.0
llama-index-readers-file>=0.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
